Cristian quiroga

+++ profesional/views.py
@@ -55,109 +55,130 @@ def indexCreateProfesional(request):
 def get_All_Profesional(request, format=None):
     user = request.user
     if user.is_authenticated:
-        profesional = Profesional.objects.all()
-        contexto = {'profesional':profesional}
-        return render(request, 'listProfesional.html', contexto)
+        if user.is_admin:
+            profesional = Profesional.objects.all()
+            contexto = {'profesional':profesional}
+            return render(request, 'administradorProfesionales.html', contexto)
+        else:
+            messages.add_message(request=request, level = messages.SUCCESS, message="No tiene suficientes permisos para ingresar a esta página")
+            return redirect('customer')
+    else:
+        return redirect('login2')
+
+def get_Profesional(request, idProfesional):
+    user = request.user
+    if user.is_authenticated:
+        if user.is_admin:
+            profesional = Profesional.objects.get(pk=idProfesional)
+            contexto = {'profesional':profesional}
+            return render(request, 'administradorProfesionales.html', contexto)
+        else:
+            messages.add_message(request=request, level = messages.SUCCESS, message="No tiene suficientes permisos para ingresar a esta página")
+            return redirect('customer')
     else:
         return redirect('login2')
 
 def addProfesional(request):
     user = request.user
     if user.is_authenticated:
-        global profesional
-        foto = request.FILES['Imagen']
-        nombre = request.POST['Nombre']
-        apellido = request.POST['Apellido']
-        curp = request.POST['CURP']
-        correo = request.POST['Correo']
-        numero_1 = request.POST['Número 1']
-        numero_2 = request.POST['Número 2']
-        redes = request.POST['Redes']
-        ubicacion = request.POST['Ubicación']
-        especialidades = request.POST['Especialidades']
-        servicios = request.POST['Servicios']
-        valor = request.POST['Valor']
-        redirectUrl = ''
-        h_list_array = Profesional.objects.order_by('nombre')
-        for h in h_list_array:
-            if curp == h.curp:
-                messages.add_message(request, messages.INFO, 'CURP ingresado ya existe')
+        if user.is_admin:
+            global profesional
+            foto = request.FILES['Imagen']
+            nombre = request.POST['Nombre']
+            if isinstance(nombre,int):
+                return Response({'Msj':'Error, el nombre debe contener letras, no números'})
+            apellido = request.POST['Apellido']
+            if isinstance(apellido,int):
+                return Response({'Msj':'Error, el apellido debe contener letras, no números'})
+            correo = request.POST['Correo']
+            numero_1 = request.POST['Número 1']
+            numero_2 = request.POST['Número 2']
+            redes = request.POST['Redes']
+            ubicacion = request.POST['Ubicación']
+            especialidades = request.POST['Especialidades']
+            servicios_valor = request.POST['Servicios y valor']
+            horario_laboral = request.POST['Horario laboral']
+            redirectUrl = ''
+            h_list_array = Profesional.objects.order_by('nombre')
+            if nombre == '' or apellido == '':
+                messages.add_message(request=request, level = messages.ERROR, message="No ha ingresado un campo requerido")
                 return render(request, 'createProfesional.html', {"profesional" : profesional})
-            elif correo == h.correo:
-                messages.add_message(request, messages.INFO, 'Correo ingresado ya existe')
+            if correo != '' and es_correo_valido(correo) == False:
+                messages.add_message(request=request, level = messages.ERROR, message="Ha ingresado un correo inválido")
                 return render(request, 'createProfesional.html', {"profesional" : profesional})
-        if nombre == '' or apellido == '' or curp == '':
-            messages.add_message(request=request, level = messages.ERROR, message="No ha ingresado un campo requerido")
-            return render(request, 'createProfesional.html', {"profesional" : profesional})
-        if correo != '' and es_correo_valido(correo) == False:
-            messages.add_message(request=request, level = messages.ERROR, message="Ha ingresado un correo inválido")
-            return render(request, 'createProfesional.html', {"profesional" : profesional})
+            else:
+                try:
+                    profesional = Profesional.objects.create(
+                        imagen_profesional = foto,
+                        nombre = nombre,
+                        apellido = apellido,
+                        correo = correo,
+                        numero_1 = numero_1,
+                        numero_2 = numero_2,
+                        redes = redes,
+                        ubicacion = ubicacion,
+                        especialidades = especialidades,
+                        servicios_valor = servicios_valor,
+                        horario_laboral = horario_laboral
+                    )
+                    messages.add_message(request=request, level = messages.SUCCESS, message="Profesional agregado correctamente")
+                    redirect('pageadmin')
+                except Exception as e:
+                    messages.add_message(request = request, level = messages.ERROR,message="Ha ocurrido un error al agregar al profesional")
+                    redirect('customer')
+            return redirect('pageadmin')
         else:
-            try:
-                profesional = Profesional.objects.create(
-                    imagen_profesional = foto,
-                    nombre = nombre,
-                    apellido = apellido,
-                    curp = curp,
-                    correo = correo,
-                    numero_1 = numero_1,
-                    numero_2 = numero_2,
-                    redes = redes,
-                    ubicacion = ubicacion,
-                    especialidades = especialidades,
-                    servicios = servicios,
-                    valor = valor
-                )
-                messages.add_message(request=request, level = messages.SUCCESS, message="Profesional agregado correctamente")
-                redirectUrl = 'pageadmin'
-            except Exception as e:
-                messages.add_message(request = request, level = messages.ERROR,message="Ha ocurrido un error al agregar al profesional")
-                redirectUrl = 'pageadmin'
-        return redirect(redirectUrl)
+            messages.add_message(request=request, level = messages.SUCCESS, message="No tiene suficientes permisos para ingresar a esta página")
+            return redirect('login2')
     else:
         return redirect('login2')
 
 def updateProfesional(request, idProfesional):
     user = request.user
     if user.is_authenticated:
-        foto = request.FILES['Imagen']
-        nombre = request.POST['Nombre']
-        apellido = request.POST['Apellido']
-        curp = request.POST['CURP']
-        correo = request.POST['Correo']
-        numero_1 = request.POST['Número 1']
-        numero_2 = request.POST['Número 2']
-        redes = request.POST['Redes']
-        ubicacion = request.POST['Ubicación']
-        especialidades = request.POST['Especialidades']
-        servicios = request.POST['Servicios']
-        valor = request.POST['Valor']
-        redirectUrl = ''
-        profesional = Profesional.objects.get(id=idProfesional)
-        if nombre == '' or apellido == '' or curp == '':
-            messages.add_message(request=request, level = messages.SUCCESS, message="El Nombre es un campo requerido")
-            return render(request, 'updateProfesional.html', {"profesional" : profesional} )
-        else:
-            try:
-                profesional.imagen_profesional = foto
-                profesional.nombre = nombre
-                profesional.apellido = apellido
-                profesional.curp = curp
-                profesional.correo = correo
-                profesional.numero_1 = numero_1
-                profesional.numero_2 = numero_2
-                profesional.redes = redes
-                profesional.ubicacion = ubicacion
-                profesional.especialidades = especialidades
-                profesional.servicios = servicios
-                profesional.valor = valor
-                profesional.save()
-                messages.add_message(request=request, level = messages.SUCCESS, message="Profesional editado correctamente")
-                return redirect('updateProfesional')
-            except Exception as e:
-                messages.add_message(request=request, level = messages.SUCCESS, message="Ha ocurrido un error al editar al profesional")
+        if user.is_admin:
+            #foto = request.FILES['Imagen']
+            nombre = request.POST['Nombre']
+            if isinstance(nombre,int):
+                return Response({'Msj':'Error, el nombre debe contener letras, no números'})
+            apellido = request.POST['Apellido']
+            if isinstance(apellido, int):
+                return Response({'Msj':'Error, el apellido debe contener letras, no números'})
+            correo = request.POST['Correo']
+            numero_1 = request.POST['Número 1']
+            numero_2 = request.POST['Número 2']
+            redes = request.POST['Redes']
+            ubicacion = request.POST['Ubicación']
+            especialidades = request.POST['Especialidades']
+            servicios_valor = request.POST['Servicios y valor']
+            redirectUrl = ''
+            profesional = Profesional.objects.get(id=idProfesional)
+            if nombre == '' or apellido == '':
+                messages.add_message(request=request, level = messages.SUCCESS, message="El Nombre es un campo requerido")
                 return render(request, 'updateProfesional.html', {"profesional" : profesional} )
-            return redirect(redirectUrl)
+            else:
+                try:
+                    #profesional.imagen_profesional = foto
+                    profesional.nombre = nombre
+                    profesional.apellido = apellido
+                    profesional.correo = correo
+                    profesional.numero_1 = numero_1
+                    profesional.numero_2 = numero_2
+                    profesional.redes = redes
+                    profesional.ubicacion = ubicacion
+                    profesional.especialidades = especialidades
+                    profesional.servicios_valor = servicios_valor
+                    profesional.horario_laboral
+                    profesional.save()
+                    messages.add_message(request=request, level = messages.SUCCESS, message="Profesional editado correctamente")
+                    return redirect('get_All_Profesional')
+                except Exception as e:
+                    messages.add_message(request=request, level = messages.SUCCESS, message="Ha ocurrido un error al editar al profesional")
+                    return render(request, 'updateProfesional.html', {"profesional" : profesional} )
+                return redirect(redirectUrl)
+        else:
+            messages.add_message(request=request, level = messages.SUCCESS, message="No tiene suficientes permisos para ingresar a esta página")
+            return redirect('customer')
     else:
         return redirect('login2')
 
@@ -165,17 +186,21 @@ def updateProfesional(request, idProfesional):
 def indexUpdateProfesional(request):
     user = request.user
     if user.is_authenticated:
-        if Profesional.objects.filter().exists():
-            firstProfesional = Profesional.objects.filter()[:1].get()
-            profesional = Profesional.objects.get(id=firstProfesional.id)
-            msg = "Profesional Configurado"
-            profesional.save()
-            return render(request, 'updateProfesional.html', {"profesional" : profesional, "msgGood":msg} )  
+        if user.is_admin:
+            if Profesional.objects.filter().exists():
+                firstProfesional = Profesional.objects.filter()[:1].get()
+                profesional = Profesional.objects.get(id=firstProfesional.id)
+                msg = "Profesional Configurado"
+                profesional.save()
+                return render(request, 'updateProfesional.html', {"profesional" : profesional, "msgGood":msg} )  
+            else:
+                profesionalprimary = Profesional.objects.create(
+                )
+                messages.add_message(request=request, level = messages.SUCCESS, message="Por favor, vuelve a ingresar al area de 'profesional'")
+                return redirect('pageadmin')
         else:
-            profesionalprimary = Profesional.objects.create(
-            )
-            messages.add_message(request=request, level = messages.SUCCESS, message="Porfavor, vuelve a ingresar al area de 'profesional'")
-            return redirect('pageadmin')
+            messages.add_message(request=request, level = messages.SUCCESS, message="No tiene suficientes permisos para ingresar a esta página")
+            return redirect('customer')
     else:
         return redirect('login2')
 
@@ -183,12 +208,24 @@ def indexUpdateProfesional(request):
 def deleteProfesional(request, idProfesional):
     user = request.user
     if user.is_authenticated:
-        if Profesional.objects.filter(id=idProfesional).exists():
-            Profesional.objects.filter(pk=idProfesional).delete()
-            messages.add_message(request=request, level = messages.SUCCESS, message="Profesional eliminado correctamente")
-            return redirect('pageadmin')
+        if user.is_admin:
+            if Profesional.objects.filter(id=idProfesional).exists():
+                Profesional.objects.filter(pk=idProfesional).delete()
+                messages.add_message(request=request, level = messages.SUCCESS, message="Profesional eliminado correctamente")
+                return redirect('pageadmin')
+            else:
+                messages.add_message(request=request, level = messages.ERROR, message="No existe el profesional")
+                return redirect('pageadmin')
         else:
-            messages.add_message(request=request, level = messages.ERROR, message="No existe el profesional")
-            return redirect('pageadmin')
+            messages.add_message(request=request, level = messages.SUCCESS, message="No tiene suficientes permisos para ingresar a esta página")
+            return redirect('customer')
     else:
         return redirect('login2')
+
+def editarProfesional(request, idProfesional):
+    profesional = Profesional.objects.get(pk=idProfesional)
+    return render(request, "updateProfesional.html", {"profesional":profesional})
+
+def detalleProfesional(request, idProfesional):
+    profesional = Profesional.objects.get(pk=idProfesional)
+    return render(request, "detalleProfesional.html", {"profesional":profesional})
